+++ tests/console/conftest.py
@@ -37,14 +37,22 @@ def mock_clone(self, source, dest):
     shutil.copytree(str(folder), str(dest))
 
 
+@pytest.fixture
+def installed():
+    return Repository()
+
+
 @pytest.fixture(autouse=True)
-def setup(mocker, installer):
+def setup(mocker, installer, installed):
     # Set Installer's installer
     p = mocker.patch('poetry.installation.installer.Installer._get_installer')
     p.return_value = installer
 
     p = mocker.patch('poetry.installation.installer.Installer._get_installed')
-    p.return_value = Repository()
+    p.return_value = installed
+
+    p = mocker.patch('poetry.repositories.installed_repository.InstalledRepository.load')
+    p.return_value = installed
 
     # Patch git module to not actually clone projects
     mocker.patch('poetry.vcs.git.Git.clone', new=mock_clone)
@@ -80,6 +88,24 @@ class Locker(BaseLocker):
         self._local_config = local_config
         self._lock_data = None
         self._content_hash = self._get_content_hash()
+        self._locked = False
+        self._lock_data = None
+
+    def is_locked(self):
+        return self._locked
+
+    def locked(self, is_locked=True):
+        self._locked = is_locked
+
+        return self
+
+    def mock_lock_data(self, data):
+        self.locked()
+
+        self._lock_data = data
+
+    def is_fresh(self):
+        return True
 
     def _write_lock_data(self, data):
         self._lock_data = None